Pin `sequelize-3.23.4`.
Unfortunately Sequelize is not following their semver promise and
they've broken the behaviour of `findAndCount` with inclusions in their
latest release (`3.23.6`). We want to update to the latest release which
includes a security fix, but the changes have a non-trivial impact in
our code, so we're pinning our Sequelize dependency to `3.23.4` which
behaves as we expected it to until we have the opportunity to
investigate further.